refactor: 폴더 구조 재구성 #13
- controller/, service/ 폴더로 분리
- mock 파일을 battles/mock/으로 이동
- import 경로 수정
- dev 브랜치 구조에 맞춰 정리
