+++ backend/src/battles/mock/battleResults.mock.ts
